+++ src/components/FixedCostsComponents/FixedCostDataInput.vue
