+++ src/components/FixedCostsComponents/FixedCostTable.vue
